+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.5.0
+
+* `JisonParser` adds two methods: `getParserErrorMessage` and `getLexerErrorMessage`. They're used for providing customized messasges, such as I18N.
+* `DefaultJisonParser` is added. It is the default base class of the generated Parser.
+* `%extends` is supported
+    * `%class` doesn't support `extends` any more
+
 # 0.3.0
 
 * Migrate to null-safety